setup: stop claiming that we are under GPL in the "license" field of the PyPI database Unfortunately, there is no way to claim that we are under a Free Software/Open Source licence without also claiming to be under a licence that we are not or claiming to have approval from DFSG or OSI, which we haven't.
Until now, I erred on the side of choosing the licence that is closest to our
from the list (GPL), but that was a bad idea and now I'm erring on the side of
not including a machine-readable licensing claim at all.

Hopefully humans who are interested will quickly find out that we are actually
under a Real Free Software Licence.

But really, this underscores that we need to talk to FSF, edit our licence for
clarity of intent, and submit it to DFSG/OSI.