setup.cfg: remove --find-links
This is only useful when running "setup.py install" directly, and is not
consulted when following our recommended "pip install .".
Platform-specific directories of wheels (that we, Tahoe, host ourselves)
are available for some platforms: these are described in
docs/INSTALL.rst and must be provided as a pip argument, not a
setup.cfg/easy_install configuration:

  pip install --find-links=https://tahoe-lafs.org/deps/

This also removes our involvement with eggs (which, I believe, are
usually downloaded insecurely by easy_install).