magic_folder_cli: refactor the join command
join takes a parameter of type Options that holds the commandline
parameters. This commit spilts join into two functions -- a wrapper
join function that still takes options and another that would be moved
closer to the magic_folder module, so that we have some sort of a
python API for every magic-folder command.