Drop PIP_NO_INDEX
There are pinned versions of pip etc in tox.ini, which isn't available
in the wheelhouse, so setting PIP_NO_INDEX=true will not work. We will
need to take some extra steps to add the required versions of those
packages to the wheelhouse, which will make CI configuration more brittle.