change exception handling
NoServersError cannot happen since we are using http API to the tahoe
client node and if that node cannot reach the tahoe grid, it would
result in a timeout error or a connection error or an unknown host
error etc. So, catch all exceptions, write a log and traceback and
continue to retry.