infrastructure to fake HTTP 504
In the event where the magic-folder is able to talk to tahoe client
but the tahoe client is unable to talk to the storage node, the http
client (magic-folder client) talking to the tahoe node would time out.
The added code creates such a Resource that always replies with a 504
code.