Handle republishing eliot logs in the face of errors.
It turns out that the state of `eliot.MemoryLogger.messages` in unpredicatble,
after `eliot.testing.check_for_errors` is called[1]. Some of the messages will
have been serialized in-place, and some will not.
In [#366](https://github.com/itamarst/eliot/issues/469), I changed the code from
supporting the case when all of the messages were serialized (which is
probably the common success case), to when none of them are (which is the
case when there are unflushed tracebacks).

This changes things around to deal with both cases.