Paramterize the synchronization integration tests on ways of snapshotting.
The existing tests explicitly request a snapshot of a given file, when
testing synchronization. Given that scanning will trigger a snapshot,
it would be nice to re-use those tests but using the scanner instead.

This adds a new parameterized test fixture that will cause a snapshot
to be taken, that tests both methods of taking a snapshot.