Ensure that there is only a single `MagicFolderConfig` instance per folder.
On windows, we can't delete a file that is open. Thus, to be able to delete a
magic-folder state database (i.e. when we leave a folder), we need to make sure
that any connections we have to it are closed.

To make this easy to ensure, we keep a cache of `MagicFolderConfig` instances
on `GlobalConfigDatabase`, so that we only have a single database connection
to the state database.