Introduce a `temp_filepath` pytest fixture returning a `FilePath`.
There are a bunch of bare `tempdir.mkdtemp` calls in the integration tests,
likely related to the confusing fact that we had a session scoped `temp_dir`
factory, in addition to the builtin `tmpdir` and `temp_path` function-scoped
fixtures.

This introduces a new function scope fixture that wraps `temp_path` returning
a `FilePath`, and adjusts the tests to use that consitently, instead of calling
`tempdir.mkdtemp` directly. This means that the directories will be cleaned up
automatically.